restructured project and modified functionality
Before this commit this project would set up a full server running ONLY
openvpn and thereby modifiying multiple parts of the system.
After this commit this project only cares about openvpn and setting it up on any server and touching only the parts that are necessary.

In detail the changes are:
- removed hardening and audit parts
- removed package upgrade and periodical update
- removed dnsmasq
- modified firewall tasks and removed server reboot
- running openvpn as root instead of openvpn user
- added multi-host support by modifiying variable loading and inventory
structure
- added possibility to set DN manually